+++ tests/testthat/test_projects_repos.R
@@ -8,22 +8,23 @@ test_that("gl_list_projects work", {
 
 # gl_list_user_projects ----
 # Chances are "testor" is one of the latest project with activity because of other unit tests
-all_user_projects <- gl_list_user_projects(user_id = test_user_id, max_page = 1, order_by = "last_activity_at")
+all_user_projects <- gl_list_user_projects(user_id = test_user_id, max_page = 2, order_by = "last_activity_at")
 
 test_that("gl_list_user_projects work", {
-  some_projects <- paste0("testor.", c("macos", "windows", "release",
-     "devel", "coverage", "main", "release.master"))
-  expect_true(all(some_projects %in% all_user_projects[["name"]]))
+  expect_true(all(test_project_name %in% all_user_projects[["name"]]))
   expect_true(all(c("id", "name", "path") %in% names(all_user_projects)))
 })
 
 # gl_list_group_projects ----
-all_group_projects <- gl_list_group_projects(group_id = test_group_id, max_page = 1)
+all_group_projects <- gl_list_group_projects(group_id = test_subgroup_id, max_page = 1)
 
 test_that("gl_list_group_projects work", {
-  some_projects <- c("publication_guide")
-  expect_true(all(some_projects %in% all_group_projects[["name"]]))
-  expect_true(all(c("id", "name", "path") %in% names(all_group_projects)))
+  if (nrow(all_group_projects) >= 1) {
+    if (test_subgroup_name == "dontdelete.subgroup.for.gitlabr") {
+      expect_true(test_subgroup_project_name %in% all_group_projects[["name"]])
+    }
+    expect_true(all(c("id", "name", "path") %in% names(all_group_projects)))
+  }
 })
 
 # gl_get_project ----
@@ -41,11 +42,23 @@ test_that("gl_proj_req works", {
 })
 
 # gl_get_project_id ----
-# Can not be really tested because gitlab.com is too big
-# except with user namespace ? No
-# gl_get_project_id(paste0(all_user_projects$namespace.path[1], "testor"), max_page = 3)
-# gitlab(req = "projects", gitlab_url = file.path(test_url, all_user_projects$namespace.path[1]), max_page = 1)
+# Can not be fully tested because gitlab.com is too big
+# => Assume that the last modified is the current project
+# because of unit tests
 
+test_that("gl_get_project_id works", {
+  expect_message(
+    {
+      the_retrieved_id <- gl_get_project_id(test_project_name,
+        max_page = 3, owned = TRUE,
+        order_by = "last_activity_at"
+      )
+    },
+    regexp = paste0("Project found: ", test_project_id)
+  )
+
+  expect_equal(as.character(the_retrieved_id), test_project_id)
+})
 # gl_archive ----
 # Dont want to test archiving project
 
@@ -56,26 +69,24 @@ test_that("gl_proj_req works", {
 #   to = "6b9d22115a93ab009d64f857dca346c0e105d64a")
 
 # test_that("Compare works", {
-#   
+#
 #   expect_s3_class(my_gitlab(compare_refs
 #                     , test_project
 #                     , "f6a96d975d9acf708560aac120ac1712a89f2a0c"
 #                     , "ea86a3a8a22b528300c03f9bcf0dc91f81db4087")
 #           , "data.frame")
-#             
+#
 # })
 
 # gl_get_commits ----
 my_commits <- gl_get_commits(test_project, ref_name = get_main())
 
 test_that("Commits work", {
-  
   my_commit <- gl_get_commits(test_project, commit_sha = my_commits$id[1])
-  
+
   expect_s3_class(my_commits, "data.frame")
   expect_s3_class(my_commit, "data.frame")
   expect_gt(length(intersect(names(my_commits), names(my_commit))), 0L)
-
 })
 
 
@@ -83,13 +94,11 @@ test_that("Commits work", {
 # gl_get_diff ----
 # The commit with CI is the last one in main branch
 the_diff <- gl_get_diff(test_project, my_commits$short_id[1])
-  
-test_that("gl_get_diff work", {
 
+test_that("gl_get_diff work", {
   expect_s3_class(the_diff, "data.frame")
   expect_equal(nrow(the_diff), 1)
-  expect_equal(the_diff$old_path, '.gitlab-ci.yml')
-  
+  expect_equal(the_diff$old_path, ".gitlab-ci.yml")
 })
 
 # gl_new_project ----
@@ -117,4 +126,3 @@ test_that("gl_edit_project work", {
 
 # gl_delete_project ----
 # Dont test delete project because this example project is needed...
-